+++ src/content/index.ts
@@ -4,3 +4,4 @@ SPDX-License-Identifier: MPL-2.0
 */
 export * from './inlineFormat';
 export * from './optionLabels';
+export * from './fieldsMapLabels';